Feature/10 skicker 도메인 개발

+++ src/main/kotlin/modeep/hear/domain/diary/vo/StickerPlacement.kt
@@ -1,8 +1,8 @@
 package modeep.hear.domain.diary.vo
 
 data class StickerPlacement(
-    val positionX: Float = 0.0f,
-    val positionY: Float = 0.0f,
-    val rotation: Float = 0.0f,
-    val scale: Float = 1.0f
+    val positionX: Double = 0.0,
+    val positionY: Double = 0.0,
+    val rotation: Double = 0.0,
+    val scale: Double = 1.0
 )
